requirements.txt
psycopg2-binary==2.9.9
resend==0.6.0
python-dotenv==1.0.0
